FIX: Actually close on stop
The `stop(doc)` method was never being called because
`DocumentRouter.__call__` is being refined in the subclass without
calling `super()`.

We could potentially fix this was calling `super()` but in one case
('start') we want the subclass to go *first* (open the file) and in
another case ('stop') we want the subclass to go *second* (close the
file). Just using `if` blocks consistently for both seems simplest.